# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EC3D-Bridge API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/process": {
      "post": {
        "summary": "Process Product",
        "operationId": "process_product_api_process_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/status/{job_id}": {
      "get": {
        "summary": "Get Status",
        "operationId": "get_status_api_status__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "summary": "List Jobs",
        "operationId": "list_jobs_api_jobs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/errors/guidance": {
      "get": {
        "summary": "Error Guidance",
        "description": "フロントエンドが error_code → ユーザー向けメッセージを引くための辞書を返す。",
        "operationId": "error_guidance_api_errors_guidance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health/detail": {
      "get": {
        "summary": "Health Detail",
        "operationId": "health_detail_health_detail_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageRef": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "type": {
            "type": "string",
            "maxLength": 20,
            "title": "Type",
            "default": "other"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ImageRef"
      },
      "ProductData": {
        "properties": {
          "product_name": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Product Name"
          },
          "source_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Source Url"
          },
          "site": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Site"
          },
          "dimensions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Dimensions"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Colors"
          },
          "materials": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Materials"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageRef"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Images"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "default": "家具"
          }
        },
        "type": "object",
        "required": [
          "product_name",
          "source_url",
          "site",
          "images"
        ],
        "title": "ProductData"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}